Inline packaged report CSS, JS, and plotly.js basic.
Drop the Python plotly dependency so offline HTML ships the smaller basic bundle from sibling assets instead of the full library.
